movd login comp, renamed auth file/folder, built out rough auth HOC
