+++ client/src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Friends from './components/Friends.js';
-import Login from './components/Login';
+import Login from './components/loginComponents/Login';
 import './App.css';
 
 function App() {
